x86/mm: widen per-domain page counters to abolish the 16TB limit
Hypercall ABI:
 - xen_domctl_shadow_op.mb is widened from uint32_t to uint64_aligned_t
   so XEN_DOMCTL_SHADOW_OP_{GET,SET}_ALLOCATION can describe shadow
   pools beyond the previous 4TB cap (uint32_t MB).

Tools:
 - xc_shadow_control(): 'mb' is now uint64_t* to match the ABI;
   readback no longer truncates the returned value.  Callers updated:
   - tools/python/xen/lowlevel/xc/xc.c (pyxc_shadow_mem_control) and
   - tools/ocaml/libs/xc/xenctrl_stubs.c (stub_shadow_allocation_{get,set}).

Truncation fixed:
 - xen/arch/arm/include/asm/p2m.h: guest_physmap_add_pages() took
   unsigned int nr_pages, which silently truncated unsigned long
   counts from common callers (notably static-shmem.c, which passes
   PFN_DOWN(psize) from an unsigned long nr_pages).  Widened to
   unsigned long.
 - xen/common/device-tree/static-memory.c: local nr_pages widened
   from unsigned int to unsigned long to match the caller chain.

non-x86:
 - device-tree: Replace UINT_MAX with ULONG_MAX in the d->max_pages
   overflow guard for symmetry with the widened field.

Suggested-by: Jan Beulich <jbeulich@suse.com>

+++ tools/include/xenctrl.h
@@ -843,7 +843,7 @@ typedef struct xen_domctl_shadow_op_stats xc_shadow_op_stats_t;
 int xc_shadow_control(xc_interface *xch,
                       uint32_t domid,
                       unsigned int sop,
-                      unsigned int *mb,
+                      uint64_t *mb,
                       unsigned int mode);
 long long xc_logdirty_control(xc_interface *xch,
                               uint32_t domid,
